feat: Modificaciones de clases request y response
