+++ src/InstrumentalHub.Application.Contracts/Features/ICategoriaService.cs
@@ -7,6 +7,7 @@ using InstrumentalHub.Domain;
 
   
 using InstrumentalHub.Dtos;
+using InstrumentalHub.Dtos.Requests;
 using Volo.Abp.Application.Services;
 
 
@@ -14,10 +15,10 @@ namespace InstrumentalHub.Features;
 
 public interface ICategoriaService : IApplicationService
 {
-    Task<List<RequestYResponse.InstrumentoCategoriaResponse>> GetListAsync();
-    Task<RequestYResponse.InstrumentoCategoriaResponse> GetAsync(Guid id);
-    Task<RequestYResponse.InstrumentoCategoriaResponse> CreateAsync(RequestYResponse.CreateInstrumentoCategoriaRequest request);
-    Task<RequestYResponse.InstrumentoCategoriaResponse> UpdateAsync(RequestYResponse.UpdateInstrumentoCategoriaRequest request);
+    Task<List<CategoriaResponse.InstrumentoCategoriaResponse>> GetListAsync();
+    Task<CategoriaResponse.InstrumentoCategoriaResponse> GetAsync(Guid id);
+    Task<CategoriaResponse.InstrumentoCategoriaResponse> CreateAsync(CategoriaRequest.CreateInstrumentoCategoriaRequest request);
+    Task<CategoriaResponse.InstrumentoCategoriaResponse> UpdateAsync(CategoriaRequest.UpdateInstrumentoCategoriaRequest request);
     Task DeleteAsync(Guid id);
